// agent-ui/src/types/os.ts
export interface Model {
  name: string
  model: string
  provider: string
}
// ...
export interface AgentDetails {
  id: string
  name?: string
  db_id?: string
  // Model
  model?: Model
}

export interface TeamDetails {
  id: string
  name?: string
  db_id?: string

  // Model
  model?: Model
}